backend/requirements.txt
flask
flask-cors
python-dotenv
openai
supabase
PyJWT
bcrypt
google-auth
google-auth-oauthlib
requests
